Further revamping of snapshot support.
Things were much more complicated than I originally imagined. We need to
provide date-stamped versions of snapshot artifacts for Maven 3, and
non-date-stamped versions of snapshot artifacts for Maven 2, and we need to
provide metadata for the date-stamped versions in a maven-metadata.xml file
inside the versioned artifact directory.

The all-versions metadata file doesn't seem necessary, so I nixed it to keep
things simple.