Exposed a Maven repository under the URL of the build.
This makes it easier to access the Maven repository from a build, and
generally make the URL strucure more consistent with the rest of
Jenkins. Concrete benefits include taking advantages of permalinks
(and thus other plugins that produce permalinks, such as the promoted
builds plugin.)

The TransientBuildActionFactory extension point was introduced in 1.459,
but in that version it is broken. 1.466 was picked as the base version
because it's the release used for LTS.