Fixing the build problem.
ClassNotFoundError: MojoNotFoundException was because Maven decided to
pick up 2.0 of maven-plugin-api (which is specified in
jenkins-maven-plugin), over 3.0 of maven-plugin-api (which is specified
in the core) because the former is shorter from the root.

I then had a Groovy <-> GMaven version issue. I fixed this by defining
Groovy as a dependency in gmaven plugin. IIRC, GMaven plugin relies on
the project to have a compile/runtime dependency on Groovy, and when
that doesn't have the right version it dies with NoSuchMethodError.

And finally GMaven complained about missing Ant, so that was added too.